requirements.txt
asgiref==3.6.0
boto3==1.26.73
botocore==1.29.129
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
defusedxml==0.7.1
Django==4.1.9
django-anymail==9.0
django-cors-headers==3.8.0
django-extra-fields==3.0.2
django-filter==22.1
django-phonenumber-field==5.2.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djongo==1.3.6
djoser==2.2.0
dnspython==2.3.0
environs==9.5.0
gunicorn==20.1.0
idna==3.4
jmespath==1.0.1
marshmallow==3.19.0
oauthlib==3.2.2
packaging==23.1
phonenumbers==8.12.31
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.6.0
pymongo==3.12.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
requests==2.30.0
requests-oauthlib==1.3.1
s3transfer==0.6.1
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.2.4
transliterate==1.10.2
ua-parser==0.16.1
urllib3==2.0.2
user-agents==2.2.0
